services/smartbrain-python/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
httpx==0.28.1
supabase==2.17.0
pandas==2.3.2
numpy==2.2.6
pandas-ta==0.3.14b0
ta==0.11.0
scikit-learn==1.7.1
xgboost==3.0.4
lightgbm==4.6.0
torch==2.8.0
stable-baselines3==2.7.0
gymnasium==1.2.1
optuna==4.4.0
langchain==0.3.27
langchain-openai==0.3.32
ccxt==4.4.59
hyperliquid-python-sdk==0.19.0
eth-account==0.13.7
hmmlearn==0.3.3
python-dotenv==1.1.1
psycopg[binary]==3.2.9
